relocate pages and standardize q-page tags
