+++ src/pages/instructor/ProblemEditor.vue
@@ -1,5 +1,7 @@
 <template>
-	<h3>Statistics</h3>
+	<q-page>
+		<h3>Problem Editor</h3>
+	</q-page>
 </template>
 
 <script lang="ts">
@@ -7,6 +9,6 @@ import { defineComponent } from 'vue';
 // import { useStore } from 'vuex';
 
 export default defineComponent({
-	name: 'Statistics'
+	name: 'ProblemEditor'
 });
 </script>
