+++ src/pages/instructor/SetDetailContainer.vue
